requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.1.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.2.0
cachelib==0.13.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cmake==4.0.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
distro==1.9.0
dlib==19.24.6
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
face_recognition_models==0.3.0
filelock==3.17.0
Flask==2.0.1
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.2
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.1
Mako==1.3.8
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
openai==0.28.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pillow==11.1.0
preshed==3.0.9
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
pytz==2025.1
qrcode==8.0
razorpay==1.4.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.23
srsly==2.5.1
sympy==1.13.1
textblob==0.19.0
thinc==8.3.4
torch==2.6.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==2.2.3
wheel==0.45.1
wrapt==1.17.2
WTForms==3.2.1
yarl==1.18.3
